refactor(game-engine): refactored the whole game engine wrapper to make it more lisible + removed some components that were used for the r-type
