+++ include/game_engine/GameEngine.hpp
@@ -295,13 +295,6 @@ namespace engine {
     void runEngineTextureMode(RenderTexture& ViewTexture);
 
     /**
-     * @brief Creates an entity without transform or render component
-     * (useful for entites that should have a behavior but not physical existence
-     * like enemy spawner etc...)
-     *
-     */
-    ecs::Entity createEntity(void);
-    /**
      * @brief Update all the systems of the engine of the entites attached to the scene
      *
      * @param sceneId Id of the scene
@@ -326,310 +319,4 @@ namespace engine {
     * See camera.getRenderTexture()
     */
     void renderTextureMode(ecs::SceneID sceneId, engine::core::CameraID cameraId);
-
-
-    std::vector<std::pair<std::type_index, std::any>> getAllComponents(ecs::Entity entity);
-
-    /**
-     * @brief Creates a cube entity with specified parameters.
-     * @param pos Position of the cube.
-     * @param width Width of the cube.
-     * @param height Height of the cube.
-     * @param length Length of the cube.
-     * @param color Color of the cube.
-     * @param toggleWire Flag to toggle wireframe mode.
-     * @param wireColor Color of the wireframe.
-     * @return The created cube entity.
-     */
-    ecs::Entity createCube(Vector3 pos, float width, float height, float length, Color color = RED, bool toggleWire = false, Color wireColor = BLACK);
-    /**
-     * @brief Creates a 3D model entity from a file.
-     * @param filename Path to the model file.
-     * @param pos Position of the model.
-     * @param color Color to apply to the model.
-     * @return The created model entity.
-     */
-    ecs::Entity createModel3D(const char *filename, Vector3 pos, Color color = WHITE);
-    /**
-     * @brief Creates a skybox entity from a file.
-     * @param filename Path to the skybox file.
-     * @param pos Position of the skybox.
-     * @param color Color to apply to the skybox.
-     * @return The created skybox entity.
-     */
-    ecs::Entity createSkybox(const char *filename, Vector3 pos, Color color = WHITE);
-
-    void destroyEntity(ecs::Entity entity);
-
-    /**
-     * @brief Creates a behavior component.
-     * @tparam T Type of the behavior.
-     * @tparam Args Variadic template arguments for the behavior's constructor.
-     * @param args Arguments to be forwarded to the behavior's constructor.
-     * @return Shared pointer to the created behavior component.
-     */
-    template<typename T, typename... Args>
-    std::shared_ptr<T> createBehavior(Args&&... args)
-    {
-        return std::make_shared<T>(std::forward<Args>(args)...);
-    }
-
-    /**
-     * @brief Attaches a behavior to an entity.
-     * @param entity The entity to attach the behavior to.
-     * @param behaviour Shared pointer to the behavior component.
-     */
-    void attachBehavior(ecs::Entity entity, std::shared_ptr<ecs::components::behaviour::Behaviour> behaviour);
-
-    /**
-     * @brief Checks if the window of the game engine is open.
-     * @return True if the window is open, false otherwise.
-     */
-    bool isWindowOpen(void);
-
-    /**
-     * @brief Returns true if a key has been pressed
-     *
-     * @param key
-     * @return true
-     * @return false
-     */
-    bool isKeyPressed(ecs::components::input::Keys key);
-
-    /**
-     * @brief Returns true if a key has been released
-     *
-     * @param key
-     * @return true
-     * @return false
-     */
-    bool isKeyReleased(ecs::components::input::Keys key);
-
-    /**
-     * @brief Returns true if a key is down
-     *
-     * @param key
-     * @return true
-     * @return false
-     */
-    bool isKeyDown(ecs::components::input::Keys key);
-
-    /**
-     * @brief Returns true if a key is up
-     *
-     * @param key
-     * @return true
-     * @return false
-     */
-    bool isKeyUp(ecs::components::input::Keys key);
-
-    /**
-     * @brief Rotates an entity.
-     * @param entity The entity to rotate.
-     * @param rotation The rotation vector.
-     */
-    void rotate(ecs::Entity entity, Vector3 rotation);
-
-    /**
-    * @brief Set the rotation of an entity.
-    * @param entity The entity to rotate.
-    * @param rotation The rotation vector.
-    */
-    void setRotation(ecs::Entity entity, Vector3 rotation);
-    /**
-     * @brief Scales an entity.
-     * @param entity The entity to scale.
-     * @param scale The scale vector.
-     */
-    void scale(ecs::Entity entity, Vector3 scale);
-
-    /**
-    * @brief Set the scale of an entity.
-    * @param entity The entity to scale.
-    * @param scale The scale vector.
-    */
-    void setScale(ecs::Entity entity, Vector3 scale);
-    /**
-     * @brief Sets the animation for an entity.
-     * @param entity The entity to set the animation for.
-     * @param filename Path to the animation file.
-     */
-    void setAnimation(ecs::Entity entity, const char *filename);
-
-    /**
-     * @brief Registers an event listener.
-     * @tparam T Type of the event.
-     * @param listener Function to handle the event.
-     */
-    template<typename T>
-    void registerListener(std::function<void(T&)> listener)
-    {
-        Engine::getInstance()->registerListener<T>(listener);
-    }
-
-    /**
-     * @brief Triggers an audio sound
-     *
-     * @param sound The sound to play
-     */
-    void triggerAudio(Sound sound);
-
-    /**
-     * @brief Plays a music
-     *
-     * @param audioPath the path to the music file
-     * @return ecs::Entity
-     */
-    ecs::Entity playMusic(const std::string &audioPath, bool looping = false);
-
-    /**
-     * @brief Stops a music
-     *
-     * @param musicSource The music source entity to which the music is attached
-     */
-    void stopMusic(ecs::Entity musicSource);
-
-    /**
-     * @brief Pauses a music
-     *
-     * @param musicSource The music source entity to which the music is attached
-     */
-    void pauseMusic(ecs::Entity musicSource);
-
-    /**
-     * @brief Resumes a paused music
-     *
-     * @param musicSource The music source entity to which the paused music is attached
-     */
-    void resumeMusic(ecs::Entity musicSource);
-
-    template<typename T>
-    void emitEvent(T &event)
-    {
-        Engine::getInstance()->emitEvent(event);
-    }
-
-    /**
-     * @brief Create a Scene object
-     *
-     * @return ecs::SceneID
-     */
-    ecs::SceneID createScene();
-
-    /**
-     * @brief Deletes a scene
-     *
-     * @param id
-     */
-    void deleteScene(ecs::SceneID id);
-
-    /**
-     * @brief Activates a scene
-     *
-     * @param id
-     */
-    void activateScene(ecs::SceneID id);
-
-    /**
-     * @brief Deactivate a scene
-     *
-     * @param id
-     */
-    void deactivateScene(ecs::SceneID id);
-
-    void pauseScene(ecs::SceneID id);
-    void resumeScene(ecs::SceneID id);
-    bool isScenePaused(ecs::SceneID id);
-
-    /**
-     * @brief Adds an entity to a scene
-     *
-     * @param entity
-     * @param sceneID
-     */
-    void addEntityToScene(ecs::Entity entity, ecs::SceneID sceneID);
-
-    /**
-     * @brief Removes an entity from a scene
-     *
-     * @param entity
-     * @param sceneID
-     */
-    void removeEntityFromScene(ecs::Entity entity, ecs::SceneID sceneID);
-
-    /**
-     * @brief Create a Camera object
-     *
-     * @param pos
-     * @param target
-     * @param up
-     * @param mode
-     * @param fov
-     * @return engine::core::EngineCamera
-     */
-    engine::core::EngineCamera createCamera(Vector3 pos = {0, 0, 0}, Vector3 target = {0, 0, 0}, Vector3 up = {0, 1, 0}, int mode = CAMERA_PERSPECTIVE, float fov = 90.0f);
-
-    /**
-     * @brief Attach a camera to a scene
-     *
-     * @param sceneID
-     * @param camera
-     */
-    void attachCamera(ecs::SceneID sceneID, engine::core::EngineCamera &camera);
-
-    /**
-     * @brief Detach a camera from a scene
-     *
-     * @param sceneID
-     * @param camera
-     */
-    void detachCamera(ecs::SceneID sceneID, engine::core::EngineCamera &camera);
-
-
-    [[nodiscard]] ecs::SceneManager& getSceneManager();
-
-
-    Matrix matrixFromFloat16(const float16& matrix);
-
-    namespace entity {
-        /**
-         * @brief Get the transform matrix of the entity.
-         * @param entity The entity to get the transform matrix from.
-         * @return The transform matrix.
-         */
-        Matrix getTransformMatrix(ecs::Entity entity);
-
-        /**
-         * @brief Get the transform of the entity.
-         * @param entity The entity to get the transform from.
-         * @return The transform.
-         */
-        Transform getTransform(ecs::Entity entity);
-
-        /**
-        * @brief Retrieves a component from an entity.
-        * @tparam T Type of the component.
-        * @param entity The entity from which the component will be retrieved.
-        * @return Reference to the component of type T.
-        */
-        template<typename T>
-        T &getComponent(ecs::Entity entity)
-        {
-            return Engine::getInstance()->getComponent<T>(entity);
-        }
-
-        /**
-         * @brief Set the transform of the entity.
-         * @param entity The entity to set the transform to.
-         * @param transMatrix The transformation matrix.
-         */
-        void setTransform(ecs::Entity entity, const Vector3 &position,
-            const Vector3 &rotation, const Vector3 &scale);
-    }
-
-    namespace math {
-          Matrix matrixFromFloat16(const float16& matrix);
-
-          void ExtractCameraViewComponents(Matrix viewMatrix, Vector3& position, Vector3& target, Vector3& up);
-    }
 }
