+++ src/game_engine/core/EventHandling.cpp
@@ -1,0 +1,40 @@
+/*
+** EPITECH PROJECT, 2024
+** game-engine
+** File description:
+** EventHandling
+*/
+
+#include "game_engine/GameEngine.hpp"
+
+namespace engine {
+
+    bool isWindowOpen(void)
+    {
+        return Engine::getInstance()->isWindowOpen();
+    }
+
+    bool isKeyPressed(ecs::components::input::Keys key)
+    {
+        auto &input = Engine::getInstance()->getSingletonComponent<ecs::components::input::Input>();
+        return input.keys[static_cast<size_t>(key)].keyPressed;
+    }
+
+    bool isKeyReleased(ecs::components::input::Keys key)
+    {
+        auto &input = Engine::getInstance()->getSingletonComponent<ecs::components::input::Input>();
+        return input.keys[static_cast<size_t>(key)].keyReleased;
+    }
+
+    bool isKeyDown(ecs::components::input::Keys key)
+    {
+        auto &input = Engine::getInstance()->getSingletonComponent<ecs::components::input::Input>();
+        return input.keys[static_cast<size_t>(key)].keyDown;
+    }
+
+    bool isKeyUp(ecs::components::input::Keys key)
+    {
+        auto &input = Engine::getInstance()->getSingletonComponent<ecs::components::input::Input>();
+        return input.keys[static_cast<size_t>(key)].keyUp;
+    }
+}
