+++ src/game_engine/core/Math.cpp
@@ -1,0 +1,50 @@
+/*
+** EPITECH PROJECT, 2024
+** game-engine
+** File description:
+** Math
+*/
+
+#include "game_engine/GameEngine.hpp"
+#include "game_engine/Math.hpp"
+
+namespace engine {
+    Matrix math::matrixFromFloat16(const float16& matrix)
+    {
+        Matrix mat;
+        mat.m0 = matrix.v[0];
+        mat.m1 = matrix.v[1];
+        mat.m2 = matrix.v[2];
+        mat.m3 = matrix.v[3];
+        mat.m4 = matrix.v[4];
+        mat.m5 = matrix.v[5];
+        mat.m6 = matrix.v[6];
+        mat.m7 = matrix.v[7];
+        mat.m8 = matrix.v[8];
+        mat.m9 = matrix.v[9];
+        mat.m10 = matrix.v[10];
+        mat.m11 = matrix.v[11];
+        mat.m12 = matrix.v[12];
+        mat.m13 = matrix.v[13];
+        mat.m14 = matrix.v[14];
+        mat.m15 = matrix.v[15];
+        return mat;
+    }
+
+    void math::ExtractCameraViewComponents(Matrix viewMatrix, Vector3& position, Vector3& target, Vector3& up)
+    {
+        // The position is the translation component of the inverted view matrix
+        position.x = viewMatrix.m3;
+        // TODO: re-add the below line, temporary fix for Main3DScene
+        // position.y = viewMatrix.m7;
+        position.z = viewMatrix.m11;
+
+        // Extracting right, up, and forward vectors from the view matrix
+        //Vector3 right = {viewMatrix.m0, viewMatrix.m4, viewMatrix.m8};
+        up = {viewMatrix.m1, viewMatrix.m5, viewMatrix.m9};
+        Vector3 forward = {viewMatrix.m2, viewMatrix.m6, viewMatrix.m10};
+
+        // Calculating the target position
+        target = Vector3Add(position, Vector3Negate(forward));
+    }
+}
