+++ src/game_engine/ecs/systems/Audio.cpp
@@ -8,6 +8,7 @@
 #include "game_engine/GameEngine.hpp"
 #include "game_engine/ecs/systems/Audio.hpp"
 #include "game_engine/ecs/components/Audio.hpp"
+#include <EntityHandling.hpp>
 
 
 #include <iostream>
